pyproject.toml
[project]
name = "museloop"
version = "0.1.0"
description = "Open-source AI agent for autonomous creative multimedia pipelines"
requires-python = ">=3.12"
license = "MIT"
readme = "README.md"
dependencies = [
    "anthropic>=0.77.0",
    "langgraph>=0.3.0",
    "langchain-anthropic>=0.4.0",
    "langchain-core>=0.3.0",
    "typer>=0.15.0",
    "rich>=13.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",
    "ffmpeg-python>=0.2.0",
    "opencv-python-headless>=4.10.0",
    "Pillow>=11.0.0",
    "structlog>=24.4.0",
    "gitpython>=3.1.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.5.0",
    "transformers>=4.48.0",
    "diffusers>=0.32.0",
]
remotion = [
    # Remotion requires Node.js installed separately (npm/npx)
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[project.scripts]
museloop = "museloop.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/museloop"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
